Refactor into Core vs Ext
