+++ lib/action/core/configuration.rb
