+++ lib/action/core/contract_validator.rb
