+++ lib/action/core/swallow_exceptions.rb
