+++ lib/action/core/top_level_around_hook.rb
