+++ lib/axn.rb
@@ -4,21 +4,19 @@ module Axn; end
 require_relative "axn/version"
 
 require "interactor"
-
 require "active_support"
 
-require_relative "action/exceptions"
-require_relative "action/logging"
-require_relative "action/configuration"
-require_relative "action/top_level_around_hook"
-require_relative "action/contract"
-require_relative "action/swallow_exceptions"
-require_relative "action/hoist_errors"
-
-require_relative "action/enqueueable"
+require_relative "action/core/exceptions"
+require_relative "action/core/logging"
+require_relative "action/core/configuration"
+require_relative "action/core/top_level_around_hook"
+require_relative "action/core/contract"
+require_relative "action/core/swallow_exceptions"
+require_relative "action/core/hoist_errors"
+require_relative "action/core/enqueueable"
 
 require_relative "axn/factory"
-require_relative "action/subactions"
+require_relative "action/ext/subactions"
 
 def Axn(callable, **kwargs)
   return callable if callable.is_a?(Class) && callable < Action
