+++ spec/single_module_focus/logging_spec.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "action/core/logging"
+
 RSpec.describe Action::Logging do
   let(:interactor) do
     build_interactor(described_class) do
